Updated browser management to automatically download and use the browser driver. Changed the name of the custom WebDriverManager class to CustomWebDriverManager to avoid conflicts, as WebDriverManager is the default name of the new dependency for driver management.
