+++ automation-tests/src/test/java/ca/bc/gov/open/ui/ContactInfoValidationChars.java
@@ -1,7 +1,6 @@
 package ca.bc.gov.open.ui;
 
 import java.time.Duration;
-import java.util.Calendar;
 
 import org.junit.After;
 import org.junit.AfterClass;
@@ -13,7 +12,7 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import ca.bc.gov.open.cto.CommonUtils;
-import ca.bc.gov.open.cto.WebDriverManager;
+import ca.bc.gov.open.cto.CustomWebDriverManager;
 
 public class ContactInfoValidationChars {
 
@@ -27,15 +26,15 @@ public class ContactInfoValidationChars {
 
 	@AfterClass
 	public static void afterClass() {
-		WebDriverManager.instance = null;
+		CustomWebDriverManager.instance = null;
 	}
 
 	@Test
 	public void test() throws Exception {
-		driver = WebDriverManager.getDriver();
-		WebDriverWait driverWait = WebDriverManager.getDriverWait();
-		WebElement element = WebDriverManager.getElement();
-		WebDriverManager.getElements();
+		driver = CustomWebDriverManager.getDriver();
+		WebDriverWait driverWait = CustomWebDriverManager.getDriverWait();
+		WebElement element = CustomWebDriverManager.getElement();
+		CustomWebDriverManager.getElements();
 
 		CommonUtils.login();
