+++ automation-tests/src/test/java/ca/bc/gov/open/ui/DisputeTicketOptionsPicker.java
@@ -17,7 +17,7 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import ca.bc.gov.open.cto.CommonUtils;
-import ca.bc.gov.open.cto.WebDriverManager;
+import ca.bc.gov.open.cto.CustomWebDriverManager;
 
 public class DisputeTicketOptionsPicker {
 
@@ -28,16 +28,16 @@ public class DisputeTicketOptionsPicker {
 	
 	  @After public void tearDown() { driver.close(); driver.quit(); }
 	  
-	  @AfterClass public static void afterClass() { WebDriverManager.instance =
+	  @AfterClass public static void afterClass() { CustomWebDriverManager.instance =
 	  null; }
 	 
 
 	@Test
 	public void test() throws Exception {
-		driver = WebDriverManager.getDriver();
-		WebDriverWait driverWait = WebDriverManager.getDriverWait();
-		WebElement element = WebDriverManager.getElement();
-		WebDriverManager.getElements();
+		driver = CustomWebDriverManager.getDriver();
+		WebDriverWait driverWait = CustomWebDriverManager.getDriverWait();
+		WebElement element = CustomWebDriverManager.getElement();
+		CustomWebDriverManager.getElements();
 
 		CommonUtils.login();
