+++ automation-tests/src/test/java/ca/bc/gov/open/ui/DisputeTicketOptionsPickerByMail.java
@@ -3,7 +3,6 @@ package ca.bc.gov.open.ui;
 import java.time.Duration;
 import java.util.Calendar;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Set;
 
 import org.junit.After;
@@ -17,7 +16,7 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import ca.bc.gov.open.cto.CommonUtils;
-import ca.bc.gov.open.cto.WebDriverManager;
+import ca.bc.gov.open.cto.CustomWebDriverManager;
 
 public class DisputeTicketOptionsPickerByMail {
 
@@ -33,15 +32,15 @@ public class DisputeTicketOptionsPickerByMail {
 
 	@AfterClass
 	public static void afterClass() {
-		WebDriverManager.instance = null;
+		CustomWebDriverManager.instance = null;
 	}
 
 	@Test
 	public void test() throws Exception {
-		driver = WebDriverManager.getDriver();
-		WebDriverWait driverWait = WebDriverManager.getDriverWait();
-		WebElement element = WebDriverManager.getElement();
-		WebDriverManager.getElements();
+		driver = CustomWebDriverManager.getDriver();
+		WebDriverWait driverWait = CustomWebDriverManager.getDriverWait();
+		WebElement element = CustomWebDriverManager.getElement();
+		CustomWebDriverManager.getElements();
 
 		CommonUtils.login();
