+++ automation-tests/src/test/java/ca/bc/gov/open/ui/DisputeTicketUploadPNG.java
@@ -1,9 +1,7 @@
 package ca.bc.gov.open.ui;
 
 import java.time.Duration;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Set;
 
 import org.junit.After;
 import org.junit.AfterClass;
@@ -16,7 +14,7 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import ca.bc.gov.open.cto.CommonUtils;
-import ca.bc.gov.open.cto.WebDriverManager;
+import ca.bc.gov.open.cto.CustomWebDriverManager;
 
 public class DisputeTicketUploadPNG {
 
@@ -25,16 +23,16 @@ public class DisputeTicketUploadPNG {
 	
 	  @After public void tearDown() { driver.close(); driver.quit(); }
 	  
-	  @AfterClass public static void afterClass() { WebDriverManager.instance =
+	  @AfterClass public static void afterClass() { CustomWebDriverManager.instance =
 	  null; }
 	 
 
 	@Test
 	public void test() throws Exception {
-		driver = WebDriverManager.getDriver();
-		WebDriverWait driverWait = WebDriverManager.getDriverWait();
-		WebElement element = WebDriverManager.getElement();
-		WebDriverManager.getElements();
+		driver = CustomWebDriverManager.getDriver();
+		WebDriverWait driverWait = CustomWebDriverManager.getDriverWait();
+		WebElement element = CustomWebDriverManager.getElement();
+		CustomWebDriverManager.getElements();
 
 		DisputeTicketUploadPNG upload = new DisputeTicketUploadPNG();
 		upload.uploadPNG(element, driverWait, driver);
