+++ automation-tests/src/test/java/ca/bc/gov/open/ui/MoreThan500CharsOnAdditionalInfoNegTest.java
@@ -12,7 +12,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import ca.bc.gov.open.cto.WebDriverManager;
+import ca.bc.gov.open.cto.CustomWebDriverManager;
 
 public class MoreThan500CharsOnAdditionalInfoNegTest {
 
@@ -26,15 +26,15 @@ public class MoreThan500CharsOnAdditionalInfoNegTest {
 
 	@AfterClass
 	public static void afterClass() {
-		WebDriverManager.instance = null;
+		CustomWebDriverManager.instance = null;
 	}
 
 	@Test
 	public void test() throws Exception {
-		driver = WebDriverManager.getDriver();
-		WebDriverWait driverWait = WebDriverManager.getDriverWait();
-		WebElement element = WebDriverManager.getElement();
-		WebDriverManager.getElements();
+		driver = CustomWebDriverManager.getDriver();
+		WebDriverWait driverWait = CustomWebDriverManager.getDriverWait();
+		WebElement element = CustomWebDriverManager.getElement();
+		CustomWebDriverManager.getElements();
 
 		DisputeTicketUploadPNG upload = new DisputeTicketUploadPNG();
 		upload.uploadPNG(element, driverWait, driver);
